List all files in the emulator system dumps as text files.
No more needing to download an emulator system image, extract it, then point
the program towards the directory of the extracted system image. All we have
to do is set the Makefile to the correct value of the system dump's SDK
version. This now includes SDK versions 14 through 20 (ICS through 'L').

Also Append flags when building with readline. Not set flags.

Conflicts:
	android-blob-utility.c