Intentionally skip emulator files that have a '#' before their names.
Certain proprietary files (such as /system/bin/rild and /system/lib/libril.so)
can technically be built from source, and therefore are in the emulator, but
a source-built file will not work with some particular devices like the LG G2.
Because the file is in the emulator, it is treated as a non-proprietary file,
and therefore ignored; meaning that it would not be printed as a *needed*
file. This allows users to put a '#' before any file that they do not want to be
ignored, and therefore will be run through the program and it will print all
dependencies (assuming that there are no other pseudo-proprietary files that
the original file depends on, as one would also have to comment out
/system/lib/libril.so for this particular example).

So opening up your SDK version's text file and turning: (without quotes)
"/system/bin/rild"
into:
"#/system/bin/rild"

will result in /system/bin/rild being marked as proprietary, and will be run
through the program accordingly, instead of being ignored and passed over.