Clean up the ifdefs.
I guess there's no harm in using a pointer for reading in input from both
fgets and readline. The variable is initialized to the default values from the
header file, and whether those values that were initialized are used or not
depend on how the makefile flags are configured; but will be disregarded in
the event that the user decided to input the variables manually. If the user
decides to use readline, they will lose 256 bytes of heap space. Oh well.