Implement function to receive input.
This single function has an #ifdef to select whether it reads from fgets, or
from readline. It is much cleaner than having a bunch of ifdefs clogging up
the main method.

Also add a function to sanitize/clean all input or possible errors that the
program wasn't able to read before (filenames/directories surrounded in
apostrophes, filenames with a space at the end of them, etc)