Do not return once we find a single hit.
With the 64-bit architectures, there are oftentimes two copies of libraries,
one for 64-bit and the other for 32-bit.

This means that there's a file under:
/system/vendor/lib/egl/libEGL_adreno.so
as well as
/system/vendor/lib64/egl/libEGL_adreno.so

so once we find a hit, do not stop there, and check under the lib64
directories, because if we fail to do this, we will likely only get one half
of the necessary libraries, as it will find a hit under lib or lib64, and not
get the other library for the other respective architecture.